requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
pydantic
psycopg2-binary
python-dotenv
pydantic-settings
chromadb
# sentence-transformers
# transformers
# torch
numpy
langchain
langchain-community
langchain-chroma
pandas
google-generativeai
langchain-google-genai